refactor(weread): separate Web UI from backend
